Add playground page
This is to allow users to learn/experiment with
bpftrace without needing to install it and run it
locally.

Powered by:
https://github.com/bpftrace/playground

The Playground component will also be used
(in follow-up PRs) to replace code snippets in the
hands-on-lab and one-liner tutorial. However,
this requires more work on the server side to make
sure most of the probes are available.